flexible config file
Added the ability to type in different feature values(need to be added in after the "graph=" line). If no value it type in for a specific feature it will be set to the default. A "READMEforConfig" has been added to explain what each line of the config file is as well as the possible features for each dataset.